Default project dir value
Currently, if you start defyne without specifying a directory, you get prompted to select a project to open or create a new project. Creating a new project causes a nil reference as the value of `dir` in `createProject()` is still undefined.
This patch uses the home directory (via `os.UserHomeDir`) as a suitable default path value to avoid a crash when the user enters a project name without specifying a directory to place the new project in.
Andy mentioned that there are bonus points for extending this patch by storing the last-opened project in persistent storage via App.Preferences.